+++ src/app/eth-extended/AlethioDataSourceFactory.ts
@@ -1,6 +1,6 @@
 import { AlethioDataSource } from "app/eth-extended/AlethioDataSource";
-import { BlockDetailsStoreFactory } from "app/eth-extended/data/block/details/BlockDetailsStoreFactory";
-import { BlockValueStoreFactory } from "app/eth-extended/data/block/value/BlockValueStoreFactory";
+import { BlockDetailsStoreFactory } from "app/shared/data/block/details/BlockDetailsStoreFactory";
+import { BlockValueStoreFactory } from "app/shared/data/block/value/BlockValueStoreFactory";
 import { TxDetailsStoreFactory } from "app/eth-extended/data/tx/details/TxDetailsStoreFactory";
 import { TokenTransferStoreFactory } from "app/eth-extended/data/token/transfer/TokenTransferStoreFactory";
 import { TxGraphStoreFactory } from "app/eth-extended/data/tx/graph/TxGraphStoreFactory";