+++ src/app/eth-extended/ethExtendedPlugin.ts
@@ -1,6 +1,6 @@
 import { IPlugin } from "plugin-api";
 import { AlethioDataSourceFactory } from "./AlethioDataSourceFactory";
-import { BlockDetailsAdapter } from "./adapter/block/BlockDetailsAdapter";
+import { BlockDetailsAdapter } from "../shared/adapter/block/BlockDetailsAdapter";
 import { EthPriceAdapter } from "./adapter/EthPriceAdapter";
 import { BlockConfirmationsAdapter } from "../shared/adapter/block/BlockConfirmationsAdapter";
 import { BlockListAdapter } from "../shared/adapter/block/BlockListAdapter";
@@ -29,10 +29,9 @@ import { accountDetailsModule } from "./module/account/accountDetails/accountDet
 import { avgTimeInPoolChartModule } from "./module/dashboard/charts/avgTimeInPoolChartModule";
 import { propagationChartModule } from "./module/dashboard/charts/propagationChartModule";
 import { BlockAvgTxTimeInPoolAdapter } from "./adapter/block/BlockAvgTxTimeInPoolAdapter";
-import { blockDetailsModule } from "./module/block/blockDetails/blockDetailsModule";
 import { unclePage } from "./page/uncle/unclePage";
 import { uncleDetailsModule } from "app/shared/module/uncle/uncleDetails/uncleDetailsModule";
-import { blockTxsModule } from "./module/block/blockTxs/blockTxsModule";
+import { blockTxsModule } from "../shared/module/block/blockTxs/blockTxsModule";
 import { txParentBlockContext } from "./context/txParentBlockContext";
 import { txParentBlockOptionalContext } from "./context/txParentBlockOptionalContext";
 import { txDetailsModule } from "./module/tx/txDetails/txDetailsModule";
@@ -41,9 +40,9 @@ import { AlethioAdapterType } from "app/shared/adapter/AlethioAdapterType";
 import { dashboardPage } from "./page/dashboard/dashboardPage";
 import { uncleByHashContextType } from "./context/uncleByHashContextType";
 import { EthExtendedPluginConfig } from "./EthExtendedPluginConfig";
-import { blockBasicModule } from "app/eth-extended/module/block/blockBasic/blockBasicModule";
-import { blockAdvancedModule } from "app/eth-extended/module/block/blockAdvanced/blockAdvancedModule";
-import { blockLogsBloomModule } from "app/eth-extended/module/block/blockLogsBloom/blockLogsBloomModule";
+import { blockBasicModule } from "app/shared/module/block/blockBasic/blockBasicModule";
+import { blockAdvancedModule } from "app/shared/module/block/blockAdvanced/blockAdvancedModule";
+import { blockLogsBloomModule } from "app/shared/module/block/blockLogsBloom/blockLogsBloomModule";
 import { txBasicModule } from "app/eth-extended/module/tx/txBasic/txBasicModule";
 import { txAdvancedModule } from "app/eth-extended/module/tx/txAdvanced/txAdvancedModule";
 
@@ -59,15 +58,13 @@ const ethExtendedPlugin: IPlugin = {
         api.addDataSource("source://aleth.io/api", dataSource);
 
         api.addDataAdapter("adapter://aleth.io/block/basic", new BlockBasicInfoAdapter(dataSource));
-        api.addDataAdapter("adapter://aleth.io/extended/block/details", new BlockDetailsAdapter(dataSource));
-        api.addDataAdapter("adapter://aleth.io/block/details", new BlockDetailsAdapter(dataSource));
+        api.addDataAdapter("adapter://aleth.io/full/block/details", new BlockDetailsAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/prices/latest", new EthPriceAdapter(dataSource, logger));
         api.addDataAdapter("adapter://aleth.io/block/confirmations",
             new BlockConfirmationsAdapter(dataSource.stores.blockStateStore));
         api.addDataAdapter("adapter://aleth.io/block-range/summary", new BlockListAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/block/latestNo",
             new LatestBlockNumberAdapter(dataSource.stores.blockStateStore));
-        api.addModuleDef("module://aleth.io/block/details", blockDetailsModule(ethSymbol));
         api.addModuleDef("module://aleth.io/block/basic", blockBasicModule);
         api.addModuleDef("module://aleth.io/block/txs", blockTxsModule(ethSymbol));
         api.addModuleDef("module://aleth.io/block/advanced", blockAdvancedModule(ethSymbol));