+++ src/app/eth-extended/module/block/blockDetails/BlockDetailsSlotType.ts
@@ -1,5 +1,0 @@
-export enum BlockDetailsSlotType {
-    Confirmations = "confirmations",
-    Txs = "txs",
-    ExtraData = "extraData"
-}