+++ src/app/eth-extended/module/block/blockTxs/txHighlight/HighlightFieldsFull.ts
@@ -1,8 +1,8 @@
-import { ITxLite } from "app/eth-extended/data/tx/ITxLite";
 import { HighlightFields } from "@alethio/explorer-ui/lib/blockTxs/txHighlight/HighlightFields";
 import { HighlightFieldKey } from "./HighlightFieldKey";
+import { ITxLiteFull } from "app/shared/data/tx/lite/ITxLiteFull";
 
-export class HighlightFieldsFull extends HighlightFields<ITxLite, HighlightFieldKey> {
+export class HighlightFieldsFull extends HighlightFields<ITxLiteFull, HighlightFieldKey> {
     constructor() {
         super();
         this.fields = [{