+++ src/app/eth-memento/module/block/blockTxs/BlockTxsStateFactory.ts
@@ -1,0 +1,14 @@
+import { GridSortingOptions } from "@alethio/ui/lib/control/grid/state/GridSortingOptions";
+import { BlockTxsState } from "app/eth-memento/module/block/blockTxs/BlockTxsState";
+import { TxsViewMode } from "@alethio/explorer-ui/lib/blockTxs/TxsViewMode";
+import { HighlightFieldsMemento } from "app/eth-memento/module/block/blockTxs/txHighlight/HighlightFieldsMemento";
+
+export class BlockTxsStateFactory {
+    create() {
+        return new BlockTxsState(
+            new TxsViewMode(),
+            new HighlightFieldsMemento(),
+            new GridSortingOptions()
+        );
+    }
+}